Align web plugin trust helper
