+++ apps/web/scripts/generate-plugin-registry.ts
@@ -26,28 +26,14 @@ import { existsSync, readFileSync, writeFileSync, mkdirSync } from "node:fs";
 import { resolve, dirname } from "node:path";
 import { createRequire } from "node:module";
 import process from "node:process";
-type PluginTrustLevel =
-  | "official"
-  | "verified"
-  | "community"
-  | "private"
-  | "unknown";
-
-const OFFICIAL_PLUGIN_SCOPES = ["@skaff", "@timonteutelink"] as const;
+import {
+  determinePluginTrustBasic,
+  type PluginTrustLevel,
+} from "@timonteutelink/skaff-lib";
 
 interface ParsedPackageSpec {
   name: string;
   version?: string;
-}
-
-function isOfficialPlugin(packageName: string): boolean {
-  return OFFICIAL_PLUGIN_SCOPES.some((scope) =>
-    packageName.startsWith(`${scope}/`),
-  );
-}
-
-function determinePluginTrustBasic(packageName: string): PluginTrustLevel {
-  return isOfficialPlugin(packageName) ? "official" : "community";
 }
 
 function parsePackageSpec(packageSpec: string): ParsedPackageSpec {
