+++ packages/skaff-lib/src/browser.ts
@@ -2,5 +2,6 @@ export { findTemplate, projectSearchPathKey } from "./utils/shared-utils";
 export * from './lib/types';
 export * from './lib/logger/types';
 export * from "./core/plugins/plugin-compatibility";
+export * from "./core/plugins/plugin-trust-basic";
 export * from "./core/plugins/plugin-types";
 export * from "./core/plugins/template-view";
