+++ packages/skaff-lib/src/core/plugins/index.ts
@@ -2,6 +2,7 @@ export * from "./plugin-loader";
 export * from "./plugin-types";
 export * from "./plugin-lifecycle";
 export * from "./plugin-compatibility";
+export * from "./plugin-trust-basic";
 export * from "./plugin-trust";
 export * from "./template-view";
 export * from "./package-spec";
